requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.5
idna==3.2
python-dotenv==0.19.0
requests==2.26.0
urllib3==1.26.6
